Separate tests only methods for JWT
