--- src/resources/jwtTemplate.json
@@ -1,6 +1,0 @@
-{
-  "iss": "Domino REST API Mocha Tests",
-  "scope": "$DATA",
-  "aud": "Domino",
-  "expSeconds": 3000
-}
